Create a generic method for creating TF tensors
This method will accept a .NET type and produce a tensor.

In the case of a scalar it allocates a single element on the managed heap to back the tensor.

In the case of single or multi-dimensional array it pins the array on the managed heap and wraps it in a tensor, allowing TF to directly interact with the managed array.

In all other cases we throw.

To produce this method I used T4 templates, which autogenerate the CS source when you edit the tt file.  See https://docs.microsoft.com/en-us/visualstudio/modeling/design-time-code-generation-by-using-t4-text-templates#creating-a-design-time-t4-text-template for more details.  We use the same approach in CoreFx and other repos for dealing with type-specific functionality breakout from a generic method.